+++ include/BrainrotRegistry.hpp
@@ -6,4 +6,5 @@ struct BrainrotRegistry {
     std::unordered_map<std::string, std::string> brainrots;
     std::unordered_map<std::string, std::string> brainrotNames; // IDs are separated from the names, for example lirili_larila would be "Lirili Larila"
     bool brainrotsRanAway;
+    bool returnBrainrotsRanAway;
 };