+++ changelog.md
@@ -1,3 +1,8 @@
+# 1.0.0-beta.3
+- 2.2081 support
+- Brainrots that have run away will return due to the GD update making it impossible to feed them.
+- Fixed issue with tiktok user brainrots running away after 48 hours instead of 24.
+
 # 1.0.0-beta.2
 - add tags